+++ src/naturerec_web/status/status_blueprint.py
@@ -2,12 +2,13 @@
 The status blueprint supplies view functions and templates for conservation status scheme management
 """
 
-from flask import Blueprint, render_template, request, redirect, session
+from flask import Blueprint, render_template, request, redirect, session, abort
 from flask_login import login_required, current_user
 from naturerec_model.logic import list_status_schemes, get_status_scheme, create_status_scheme, update_status_scheme, \
     delete_status_scheme
 from naturerec_model.logic import create_status_rating, update_status_rating, delete_status_rating
 from naturerec_model.data_exchange import StatusImportHelper
+from naturerec_web.auth import requires_roles, has_roles
 from naturerec_web.request_utils import get_posted_int
 
 status_bp = Blueprint("status", __name__, template_folder='templates')
@@ -60,6 +61,7 @@ def _render_ratings_import_page(error):
 
 @status_bp.route("/list", methods=["GET", "POST"])
 @login_required
+@requires_roles(["Administrator", "Reporter", "Reader"])
 def list_all():
     """
     Show the page that lists all conservation status schemes and is the entry point for adding new ones
@@ -69,9 +71,12 @@ def list_all():
     error = None
     if request.method == "POST":
         try:
-            delete_record_id = get_posted_int("delete_record_id")
-            if delete_record_id:
-                delete_status_scheme(delete_record_id)
+            if has_roles("[Administrator]"):
+                delete_record_id = get_posted_int("delete_record_id")
+                if delete_record_id:
+                    delete_status_scheme(delete_record_id)
+            else:
+                abort(401)
         except ValueError as e:
             error = e
 
@@ -86,6 +91,7 @@ def list_all():
 @status_bp.route("/edit", defaults={"status_scheme_id": None}, methods=["GET", "POST"])
 @status_bp.route("/edit/<int:status_scheme_id>", methods=["GET", "POST"])
 @login_required
+@requires_roles(["Administrator"])
 def edit_scheme(status_scheme_id):
     """
     Serve the page to add new conservation status rating or edit an existing one and handle the appropriate action
@@ -114,9 +120,10 @@ def edit_scheme(status_scheme_id):
 @status_bp.route("/add_rating/<int:status_scheme_id>", defaults={"status_rating_id": None}, methods=["GET", "POST"])
 @status_bp.route("/edit_rating/<int:status_scheme_id>/<int:status_rating_id>", methods=["GET", "POST"])
 @login_required
+@requires_roles(["Administrator", "Reporter", "Reader"])
 def edit_rating(status_scheme_id, status_rating_id):
     """
-    Serve the page to add new species or edit an existing one and handle the appropriate action
+    Serve the page to add new rating or edit an existing one and handle the appropriate action
     when the form is submitted
 
     :param status_scheme_id: ID for the parent conservation status scheme
@@ -138,6 +145,7 @@ def edit_rating(status_scheme_id, status_rating_id):
 
 @status_bp.route("/import", methods=["GET", "POST"])
 @login_required
+@requires_roles(["Administrator"])
 def import_ratings():
     """
     Serve the page to import status ratings and handle the import when the form is submitted
